Set up station_status_env.ini to handle calls for wreck fields.  Every station was given a the same, standard wreck field.
Additionally, the ps debris field was set to display no objects, as for some reason the new wreck code was crowding the old debris around the ship very closely.  We no longer use the debris fields so turning it off should not matter.  This will also solve the problem of the perpetual debris field around the shipyard station (which is likely hardcoded)