A directory for keeping generated header files in.
You can clean it out with 'make clean-headers' from the usual place.

The main reason for having this here is so we can run:

diff -ur include_pkg include_tmp

which will show up any changed header files and hopefully make it
harder to forget to commit changed headers to svn.

I was hopeing to automate this a bit more, but the $Header$ stuff
confuses it.