Add a python script to generate dependancies automaticly.
If you have $EPIC_BASE set it will work out which files it
needs to read and just do it, otherwise run it with a list
of directories as args - it will scan them for .pog and .h
files.

Note that the output includes per file build rules -
unfortunatly the pog compiler spits out .pkg files in
lower case so if you're on a system with a case sensative
filesystem (which my laptop is) then FileName.pog and
filename.pkg don't match so make never realises that the 
targets have been built.

Also the build rules and dependancies are for every .pog
file that the script finds - which means that there may
be build rules for .pkg files that are NOT in the
$(PACKAGES) vairable in the Makefile. These .pkg's will
not be built - only files in $(PACKAGES) will be.

The output of the script goes to stdout, so redirect it
to MS_Mod_Main/packages/pogdepend