This is a big update.  Essentially it includes changes that allow the game to be played.  Previous to this most of our features were not accessible because there were no NPCs to be hired.  NPCs carry out many of the functions/features of the game.  There is a lot in here, so I will try to summarize as best I can.
First, a new, base FreeTrader operation was built. It includes much more robust handling of fight/flee functions, better fight strategies, an new approach for scanning for attacks/distress, and the ability to use Lpoints and plan routes to new systems.  Additionally the operation use the Max_Fleets number in metagame_production as a maximum value, then randomizes its size.  This is saved and restored on saves.  The operation also uses the lpoints properly, joining the cue, forming up after jumping etc.  Amazingly, the operation continues to exist and operate in the new system, which allows battles and chases between systems.  Maybe.

This new base operation was then plugged back into, and used as the base of, the following operations:  FreeTanker, FreeScience, FreeMiner, FreeTrader and Merc.  These operations now provide all the hirable npcs the player will need.  The base operation was also put back into the Trader and Guard ops, though the Guard ops could use some cleanup as much of the code was ignored, owing to their simple function.  The trader op is the most complex by far, and features many behaviours and states.

Operations were set not to cull for up to 4 hours after meeting the player.  The purpose was to give the systems the player was in a more dynamic appearance.  The switching on and off of operations simply wasn't enough to give a playable, believable field of traffic.  Switching the operations to live 4 hours after leaving the 200k mark around the player opened up a big can of worms, mainly in the way they were set to be culled, saved, and how they interacted with the games mission time and actual time.  Needless to say the solutions are all tied up in the files included.

The adjustments to combat necessitated changes to the tspilotsetup.  Through some tricky coding I managed to give the ships the ability to have multiple, renewed reactions to attacks.  This is something we never had before and GT struggled with.  Essentially what it boils down to is an attacked ship that choses to flee doesn't need to do so forever.  It can recover from this state, and go back about its business without having to be culled and restored.

I also made the freetrader type operations go hostile, and permanently remove their offers for hire after being attacked by the player. This just makes sense.  Previously the player could attack an operation, and the next time he met it could actually hire the same ship.  He could even hire ships in the middle of battle.  Obviously not good, but fixed now.

tsFleetOpPhysical was not restoring anyting from saves, and I still can't figure out exactly why.  However, I fixed it with a bit of a kludge and it works alright now.  I believe Joco had plans to make it more robust, with perhaps even the ability to save out ships locations, speed etc.  For now though it just restores ships to their init state.  Could use some looking into.

Changes were to made to tsCapsuleJump to help operations use lpoints.

tsMain and tsTime were changed to help give a much more accurate count of time between saves and during gameplay.

tsPlayerWatch had some things added for a planned HyperTime effect in LDS and capsule space, but it is commented out for now.  

tsShipComms and tsShipHire received functions to remove the comms ability from ships, in the case that the player attacks them

tsShipEscort fixed a spelling error for CapitalShips

Various headers were updated accordingly