For tsPcGuards.pog I added a check for 'liked' ships, meaning ships that are very friendly or allied, so that these can be added to the fleet calculation when facing an enemy with several allied operations nearby.  Without this multiple operations cannot work in unison and will all flee a small enemy, even if together they are much stronger.  This also is a direct opposite of the attacker's calculation for fleet strength.
tsPcFreeMiner.pog, tsPcFreeScience.pog, tsPcFreeTanker.pog and tsPcMerc.pog all had a short snippet of code added that turns off the ability to the hire a ship that is engaging in docking functions.  This prevents the ship being commanded by both the player and the operation after the ship has been hired away by the player.  This dealt with issue 181.