This update deals with issue 179.  Ships do not notice attacks from friendly or allied ships.  So this meant that the player could pound on allied factions with no fear of retribution.  I have added the WatchForAttacks function into the player_watch task so that we can notice when the player or his wingmen are doing this and quickly shift the allied or friendly factions relation down to a level that will respond to the attacks.  We won't have to worry about this for any other operations, as they won't harass each other, and the lack of reactions was probably built in on purpose to prevent ships of the same faction attacking each other because of stray shots.
I also had to fix tsTime, as it the GetsTime function was not working for CheckForAttacks.  I had adjusted it for controlling saving and resaving of factions, and proper tracking of days passed.  My fix hopefully works with both.  

tsPcMer.pog also had a fix, changing ships from cargo setups to escort setups.