This is the final update before the actual Beta
Moved the spoiler map so it wouldn't be a spoiler.

Adjusted player_start, player_start_choice, adjusted for the new StartScreen

Adjusted all the configs to set communications initiation to 'v' from 'ctrl c'.  Ctrl c would often cause a jump in the current target if the player just pressed 'c'

Added gui images for the hud overlays.  

Deactivated all the debug features in the uniGUI menu via MS_Mod_Debug.ini

Adjusted numerous html files to reflect current changes, keybinds etc.

Changed the name of menu items in MS_Mod_Main.ini

Added RunCheck to tsClaimGui, to enable turning it on and off as an menu item dependent on whether the player is near a station or not.

Edited tsExitGui to close the game on exit.  I am not sure why GT didn't do this, as allowing the player to get back to the menu and start a new game, will give them a corrupted start.

Added RunCheck to tsFactionGui, so players can only access it near stations.

Adjusted tsMain to for additions needed for the new start screen.

Added StartDolly, which adds a iDirector cycling movie to the start screen.

Made tsStationAi check that the player is actually docked to the station, and not just docked to something.

Changed the viewtype in tsViewDock for some reason.

Added strings for pod loading and unloading in crew_text

Added station AI comments to wing_text.csv

Added universal ports to the freight ships.  They are overlapping the Unload ports, so this may be bad.

Updated tsClaimGui and tsStartGui to added functions

Changed installer images

Adjusted epic_install_maker_full_complete