This update is hopefully restoring the make functions and clearing out the 'Rejected World' bug that was introduced by a call to tsPlayerSave.  The bug was fixed by adding the called function to both the 'Provides' area and as a prototype.
I also switched out all the headers for the auto generated ones.  It looks like this was not done for a long time, and there were some serious discrepancies developing.  I don't know if this will be a good thing, or a total disaster.  When doing this I also noticed that a lot of hand made headers had tasks listed with a 'task' prefix, but when done automatically, the name of the task is prefixed rather than 'task'.  There were also a few missing functions in some of the hand made headers.  We will see how things go with this.